Fix circular import preventing allowlist.py script execution
The allowlist.py script failed with ImportError due to circular dependency.

Created packit_service/worker/helpers/testing_farm_mixin.py to separate
Testing Farm-specific mixins from general handler mixins.

Extracted get_check_name_prefix() from FedoraCIJobHandler to
packit_service/utils.py as a standalone utility function.

Updated imports across testing_farm.py, checker/testing_farm.py, and
handlers modules to use the new locations.
