+++ packit_service/worker/handlers/mixin.py
@@ -35,10 +35,6 @@ from packit_service.worker.handlers.abstract import CeleryTask
 from packit_service.worker.helpers.build.copr_build import CoprBuildJobHelper
 from packit_service.worker.helpers.build.koji_build import KojiBuildJobHelper
 from packit_service.worker.helpers.sidetag import Sidetag, SidetagHelper
-from packit_service.worker.helpers.testing_farm import (
-    DownstreamTestingFarmJobHelper,
-    TestingFarmJobHelper,
-)
 from packit_service.worker.mixin import Config, ConfigFromEventMixin, GetBranches
 from packit_service.worker.monitoring import Pushgateway
 
@@ -606,68 +602,6 @@ class GetCoprBuildJobHelperForIdMixin(
                 build_targets_override=build_targets_override,
             )
         return self._copr_build_helper
-
-
-class GetTestingFarmJobHelper(Protocol):
-    package_config: PackageConfig
-    job_config: JobConfig
-    celery_task: Optional[CeleryTask] = None
-
-    @property
-    @abstractmethod
-    def testing_farm_job_helper(self) -> TestingFarmJobHelper: ...
-
-
-class GetTestingFarmJobHelperMixin(
-    GetTestingFarmJobHelper,
-    GetCoprBuildMixin,
-    ConfigFromEventMixin,
-):
-    _testing_farm_job_helper: Optional[TestingFarmJobHelper] = None
-
-    @property
-    def testing_farm_job_helper(self) -> TestingFarmJobHelper:
-        if not self._testing_farm_job_helper:
-            self._testing_farm_job_helper = TestingFarmJobHelper(
-                service_config=self.service_config,
-                package_config=self.package_config,
-                project=self.project,
-                metadata=self.data,
-                db_project_event=self.db_project_event,
-                job_config=self.job_config,
-                build_targets_override=self.data.build_targets_override,
-                tests_targets_override=self.data.tests_targets_override,
-                celery_task=self.celery_task,
-            )
-        return self._testing_farm_job_helper
-
-
-class GetDownstreamTestingFarmJobHelper(Protocol):
-    celery_task: Optional[CeleryTask] = None
-
-    @property
-    @abstractmethod
-    def downstream_testing_farm_job_helper(self) -> DownstreamTestingFarmJobHelper: ...
-
-
-class GetDownstreamTestingFarmJobHelperMixin(
-    GetDownstreamTestingFarmJobHelper,
-    GetKojiBuildFromTaskOrPullRequestMixin,
-    ConfigFromEventMixin,
-):
-    _downstream_testing_farm_job_helper: Optional[DownstreamTestingFarmJobHelper] = None
-
-    @property
-    def downstream_testing_farm_job_helper(self) -> DownstreamTestingFarmJobHelper:
-        if not self._downstream_testing_farm_job_helper:
-            self._downstream_testing_farm_job_helper = DownstreamTestingFarmJobHelper(
-                service_config=self.service_config,
-                project=self.project,
-                metadata=self.data,
-                koji_build=self.koji_build,
-                celery_task=self.celery_task,
-            )
-        return self._downstream_testing_farm_job_helper
 
 
 class GetGithubCommentEvent(Protocol):
