+++ packit_service/worker/helpers/testing_farm.py
@@ -39,9 +39,12 @@ from packit_service.models import (
 )
 from packit_service.sentry_integration import send_to_sentry
 from packit_service.service.urls import get_testing_farm_info_url
-from packit_service.utils import get_package_nvrs, get_packit_commands_from_comment
+from packit_service.utils import (
+    get_check_name_prefix,
+    get_package_nvrs,
+    get_packit_commands_from_comment,
+)
 from packit_service.worker.celery_task import CeleryTask
-from packit_service.worker.handlers.abstract import FedoraCIJobHandler
 from packit_service.worker.helpers.build import CoprBuildJobHelper
 from packit_service.worker.helpers.fedora_ci import FedoraCIHelper
 from packit_service.worker.helpers.testing_farm_client import TestingFarmClient
@@ -1394,7 +1397,7 @@ class DownstreamTestingFarmJobHelper:
     @staticmethod
     def get_check_name_from_config(test_name: str, service_config: ServiceConfig) -> str:
         """Static version for use in class methods."""
-        prefix = FedoraCIJobHandler.get_check_name_prefix(service_config)
+        prefix = get_check_name_prefix(service_config)
         return f"{prefix} - {test_name}"
 
     def report(
